requirements.txt
Flask==2.3.3
reportlab==4.0.7
Pillow==10.3.0
